Re-organize the test folder, improve the code structures.
