--- test/SimpleDAppNoMain1.java
@@ -1,19 +1,0 @@
-import org.aion.abigenerator.Callable;
-
-public class SimpleDAppNoMain1 {
-
-    @Callable()
-    public static boolean test1(boolean b) {
-        return true;
-    }
-
-    @Callable()
-    public static boolean test2(int i, String s, long[] l) {
-        return true;
-    }
-
-    @Deprecated
-    public static boolean test3(int i, String s, long[] l) {
-        return true;
-    }
-}
