+++ test/org/aion/abigenerator/RandomArgumentsGenerator.java
@@ -1,3 +1,5 @@
+package org.aion.abigenerator;
+
 import org.aion.avm.abi.internal.ABICodec;
 import org.junit.Assert;
