+++ test/org/aion/abigenerator/RandomTest.java
@@ -1,4 +1,10 @@
-import org.aion.abigenerator.ABICompiler;
+package org.aion.abigenerator;
+
+import java.io.ByteArrayInputStream;
+import java.math.BigInteger;
+import java.util.Arrays;
+import java.util.HashMap;
+
 import org.aion.avm.api.ABIDecoder;
 import org.aion.avm.api.ABIEncoder;
 import org.aion.avm.api.Address;
@@ -15,10 +21,6 @@ import org.objectweb.asm.AnnotationVisitor;
 import org.objectweb.asm.ClassWriter;
 import org.objectweb.asm.Label;
 import org.objectweb.asm.MethodVisitor;
-
-import java.io.ByteArrayInputStream;
-import java.math.BigInteger;
-import java.util.HashMap;
 
 import static org.junit.Assert.assertEquals;
 import static org.objectweb.asm.Opcodes.*;
