+++ test/org/aion/abigenerator/SimpleDAppNoMain1.java
@@ -1,11 +1,6 @@
-import org.aion.abigenerator.Callable;
-import org.aion.avm.api.ABIDecoder;
-import org.aion.avm.api.BlockchainRuntime;
+package org.aion.abigenerator;
 
-public class SimpleDApp {
-    public static byte[] main() {
-        return ABIDecoder.decodeAndRunWithClass(SimpleDApp.class, BlockchainRuntime.getData());
-    }
+public class SimpleDAppNoMain1 {
 
     @Callable()
     public static boolean test1(boolean b) {
