+++ test/org/aion/abigenerator/SimpleDAppWithMain.java
@@ -1,10 +1,11 @@
-import org.aion.abigenerator.Callable;
+package org.aion.abigenerator;
+
 import org.aion.avm.api.ABIDecoder;
 import org.aion.avm.api.BlockchainRuntime;
 
-public class SimpleDAppWrongCallable {
+public class SimpleDAppWithMain {
     public static byte[] main() {
-        return ABIDecoder.decodeAndRunWithClass(SimpleDApp.class, BlockchainRuntime.getData());
+        return ABIDecoder.decodeAndRunWithClass(SimpleDAppWithMain.class, BlockchainRuntime.getData());
     }
 
     @Callable()
@@ -21,10 +22,4 @@ public class SimpleDAppWrongCallable {
     public static boolean test3(int i, String s, long[] l) {
         return true;
     }
-
-    @Callable()
-    protected boolean test4(int i, String s, long[] l) {
-        return true;
-    }
 }
-
