+++ test/org/aion/abigenerator/Type.java
@@ -1,3 +1,5 @@
+package org.aion.abigenerator;
+
 public enum Type {
     BYTE            (1),
     BOOLEAN         (2),
